move introduced functionality to webpack_loader.contrib.pages
