+++ webpack_loader/templatetags/webpack_loader.py
@@ -1,12 +1,7 @@
-import functools
-
 from django import template, VERSION
-from django.contrib.staticfiles import finders
 from django.utils.safestring import mark_safe
 
 from .. import utils
-from ..config import load_config
-from ..utils import get_unique_entrypoint_files
 
 register = template.Library()
 
@@ -46,56 +41,3 @@ def get_files(bundle_name, extension=None, config="DEFAULT"):
     :return: a list of matching chunks
     """
     return utils.get_files(bundle_name, extension=extension, config=config)
-
-
-@register.simple_tag(takes_context=True)
-def register_entrypoint(context, entrypoint):
-    if not hasattr(context, "webpack_entrypoints"):
-        context.webpack_entrypoints = []
-    context.webpack_entrypoints.insert(0, entrypoint)
-
-
-@register.simple_tag(takes_context=True)
-def render_css(context, config="DEFAULT"):
-    """Render <style> and/or <link> tags, depending on the use of CRITICAL_CSS. Should be put in the <head>"""
-    entrypoints = getattr(context, "webpack_entrypoints", [])
-    preloadTags = []
-    noscriptTags = []
-    for file in get_unique_entrypoint_files(entrypoints, "css", config):
-        preloadTags.append(f'<link rel="preload" href="{file["url"]}" as="style" '
-                           f'onload="this.onload=null;this.rel=\'stylesheet\'">')
-        noscriptTags.append(f'<link rel="stylesheet" href="{file["url"]}">')
-    criticalPath = finders.find(f"bundles/{entrypoints[-1]}.critical.css")
-    cfg = load_config(config)
-    if context["request"].first_visit and cfg["CRITICAL_CSS_ENABLED"] and criticalPath:
-        with open(criticalPath) as f:
-            criticalCss = f.read()
-        return mark_safe(
-            f"<style>{criticalCss}</style>\n"
-            f"{''.join(preloadTags)}\n"
-            f"<script>{inline_static_file('bundles/cssrelpreload.js')}</script>\n"
-            f"<noscript>{''.join(noscriptTags)}</noscript>"
-        )
-    else:
-        return mark_safe("".join(noscriptTags))
-
-
-@register.simple_tag(takes_context=True)
-def render_js(context, config="DEFAULT"):
-    files = get_unique_entrypoint_files(getattr(context, "webpack_entrypoints", []), "js", config)
-    return mark_safe("".join(f"<script src='{file['url']}'></script>" for file in files))
-
-
-@functools.lru_cache
-def inline_static_file(path):
-    with open(finders.find(path)) as f:
-        return mark_safe(f.read())
-
-
-@functools.lru_cache
-def inline_entrypoint(entrypoint, extension, config="DEFAULT"):
-    inlined = ""
-    for file in get_unique_entrypoint_files((entrypoint,), extension, config=config):
-        with open(finders.find(f"bundles/{file['name']}")) as f:
-            inlined += f.read()
-    return mark_safe(inlined)
